+++ libs/utils/decorate.js
@@ -1,42 +1,6 @@
 // Shared block decorate functions
 
 /**
- * Decorates buttons within a given element by adding
- * appropriate classes based on their parent elements.
- *
- * The function searches for buttons within the provided element that
- * match the selectors 'em a', 'strong a', and 'p > a strong'.
- * It then assigns classes to these buttons based on their parent elements
- * and any custom classes found in the button text.
- *
- * @param {HTMLElement} el - The element within which to search for and decorate buttons.
- */
-export function decorateButtons(el) {
-  const buttons = el.querySelectorAll('em a, strong a, p > a strong');
-  if (buttons.length === 0) return;
-  const buttonTypeMap = { STRONG: 'primary', EM: 'secondary', A: 'link' };
-  buttons.forEach((button) => {
-    let target = button;
-    const parent = button.parentElement;
-    const buttonType = buttonTypeMap[parent.nodeName] || 'primary';
-    if (button.nodeName === 'STRONG') {
-      target = parent;
-    } else {
-      parent.insertAdjacentElement('afterend', button);
-      parent.remove();
-    }
-    target.classList.add('button', buttonType);
-    const customClasses = target.textContent && [...target.textContent.matchAll(/#_button-([a-zA-Z-]+)/g)];
-    if (customClasses) {
-      customClasses.forEach((match) => {
-        target.textContent = target.textContent.replace(match[0], '');
-        target.classList.add(match[1]);
-      });
-    }
-  });
-}
-
-/**
  * Adjusts the focal point of an image within a picture element
  * based on the provided child element's content.
  *
